move batch_helper to prompt_target as it depends on that interface
