+++ pyrit/orchestrator/skeleton_key_orchestrator.py
@@ -8,7 +8,7 @@ from uuid import uuid4
 
 from colorama import Fore, Style
 
-from pyrit.common.batch_helper import batch_task_async
+from pyrit.prompt_target.batch_helper import batch_task_async
 from pyrit.common.path import DATASETS_PATH
 from pyrit.models import (
     PromptRequestResponse,
